+++ src/test/java/com/ojodev/cookinghero/recipes/RecipesApplicationTest.java
@@ -7,7 +7,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class RecipesApplicationTests {
+public class RecipesApplicationTest {
 
 	@Test
 	public void contextLoads() {
