+++ src/test/java/com/ojodev/cookinghero/recipes/api/CuisineTypesApiControllerGetAllTest.java
@@ -16,7 +16,8 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.util.Optional;
+import java.util.ArrayList;
+import java.util.Arrays;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.mockito.ArgumentMatchers.any;
@@ -28,7 +29,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @AutoConfigureMockMvc
-public class CuisineTypesApiControllerGetCuisineTypeTests {
+public class CuisineTypesApiControllerGetAllTest {
 
     @Autowired
     private MockMvc mvc;
@@ -43,51 +44,68 @@ public class CuisineTypesApiControllerGetCuisineTypeTests {
     private static final String LOCALE_SPANISH = "es";
 
     private static final String INVALID_LANGUAGE = "xx";
-    private static final String INVALID_ID = "xxxxx";
+    private static final String INVALID_NAME = "xxxxx";
 
     @Test
-    public void getCuisineType() throws Exception {
+    public void getAllCuisineTypes() throws Exception {
 
-        when(this.cuisineTypesBusiness.getCuisineType(any(), any())).thenReturn(Optional.of(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH));
+        when(this.cuisineTypesBusiness.getCuisineTypes(any(), any())).thenReturn(Arrays.asList(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH, CuisineTypesExamples.CUISINE_TYPE_BO_02_ENGLISH, CuisineTypesExamples.CUISINE_TYPE_BO_03_ENGLISH));
 
-        this.mvc.perform(get("/cuisine-types/{cuisine-type}", CuisineTypesExamples.CUISINE_TYPE_01_ID)
+        this.mvc.perform(get("/cuisine-types")
                 .header(HttpHeaders.ACCEPT_LANGUAGE, LOCALE_ENGLISH)
                 .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(header().string(HttpHeaders.CONTENT_LANGUAGE, LOCALE_ENGLISH))
-                .andExpect(jsonPath("$.id", is(CuisineTypesExamples.CUISINE_TYPE_01_ID)))
-                .andExpect(jsonPath("$.name", is(CuisineTypesExamples.CUISINE_TYPE_01_NAME_ENGLISH)));
+                .andExpect(jsonPath("$.length()", is(3)))
+                .andExpect(jsonPath("$[0].id", is(CuisineTypesExamples.CUISINE_TYPE_01_ID)))
+                .andExpect(jsonPath("$[0].name", is(CuisineTypesExamples.CUISINE_TYPE_01_NAME_ENGLISH)))
+                .andExpect(jsonPath("$[1].id", is(CuisineTypesExamples.CUISINE_TYPE_02_ID)))
+                .andExpect(jsonPath("$[1].name", is(CuisineTypesExamples.CUISINE_TYPE_02_NAME_ENGLISH)))
+                .andExpect(jsonPath("$[2].id", is(CuisineTypesExamples.CUISINE_TYPE_03_ID)))
+                .andExpect(jsonPath("$[2].name", is(CuisineTypesExamples.CUISINE_TYPE_03_NAME_ENGLISH)));
+    }
+
+
+    @Test
+    public void getAllCuisineTypesByName() throws Exception {
+
+        when(this.cuisineTypesBusiness.getCuisineTypes(any(), any())).thenReturn(Arrays.asList(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH, CuisineTypesExamples.CUISINE_TYPE_BO_02_ENGLISH, CuisineTypesExamples.CUISINE_TYPE_BO_03_ENGLISH));
+        when(this.cuisineTypesBusiness.getCuisineTypes(CuisineTypesExamples.CUISINE_TYPE_01_NAME_ENGLISH, LanguageEnumBO.EN)).thenReturn(Arrays.asList(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH));
+
+        this.mvc.perform(get("/cuisine-types")
+                .header(HttpHeaders.ACCEPT_LANGUAGE, LOCALE_ENGLISH)
+                .accept(MediaType.APPLICATION_JSON)
+                .param("name", CuisineTypesExamples.CUISINE_TYPE_01_NAME_ENGLISH))
+                .andExpect(status().isOk())
+                .andExpect(header().string(HttpHeaders.CONTENT_LANGUAGE, LOCALE_ENGLISH))
+                .andExpect(jsonPath("$.length()", is(1)))
+                .andExpect(jsonPath("$[0].id", is(CuisineTypesExamples.CUISINE_TYPE_01_ID)))
+                .andExpect(jsonPath("$[0].name", is(CuisineTypesExamples.CUISINE_TYPE_01_NAME_ENGLISH)));
     }
 
     @Test
-    public void getCuisineTypeDifferentLanguages() throws Exception {
+    public void getAllCuisineTypesDifferentLanguages() throws Exception {
 
-        when(this.cuisineTypesBusiness.getCuisineType(CuisineTypesExamples.CUISINE_TYPE_01_ID, LanguageEnumBO.EN)).thenReturn(Optional.of(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH));
-        when(this.cuisineTypesBusiness.getCuisineType(CuisineTypesExamples.CUISINE_TYPE_01_ID, LanguageEnumBO.ES)).thenReturn(Optional.of(CuisineTypesExamples.CUISINE_TYPE_BO_01_SPANISH));
+        when(this.cuisineTypesBusiness.getCuisineTypes(any(), eq(LanguageEnumBO.EN))).thenReturn(Arrays.asList(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH));
+        when(this.cuisineTypesBusiness.getCuisineTypes(any(), eq(LanguageEnumBO.ES))).thenReturn(Arrays.asList(CuisineTypesExamples.CUISINE_TYPE_BO_01_SPANISH));
 
-        this.mvc.perform(get("/cuisine-types/{cuisine-type}", CuisineTypesExamples.CUISINE_TYPE_01_ID)
-                .header(HttpHeaders.ACCEPT_LANGUAGE, LOCALE_ENGLISH)
-                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andExpect(header().string(HttpHeaders.CONTENT_LANGUAGE, LOCALE_ENGLISH))
-                .andExpect(jsonPath("$.id", is(CuisineTypesExamples.CUISINE_TYPE_01_ID)))
-                .andExpect(jsonPath("$.name", is(CuisineTypesExamples.CUISINE_TYPE_01_NAME_ENGLISH)));
-
-        this.mvc.perform(get("/cuisine-types/{cuisine-type}", CuisineTypesExamples.CUISINE_TYPE_01_ID)
+        this.mvc.perform(get("/cuisine-types")
                 .header(HttpHeaders.ACCEPT_LANGUAGE, LOCALE_SPANISH)
-                .accept(MediaType.APPLICATION_JSON))
+                .accept(MediaType.APPLICATION_JSON)
+                .param("name", CuisineTypesExamples.CUISINE_TYPE_01_NAME_SPANISH))
                 .andExpect(status().isOk())
                 .andExpect(header().string(HttpHeaders.CONTENT_LANGUAGE, LOCALE_SPANISH))
-                .andExpect(jsonPath("$.id", is(CuisineTypesExamples.CUISINE_TYPE_01_ID)))
-                .andExpect(jsonPath("$.name", is(CuisineTypesExamples.CUISINE_TYPE_01_NAME_SPANISH)));
+                .andExpect(jsonPath("$.length()", is(1)))
+                .andExpect(jsonPath("$[0].id", is(CuisineTypesExamples.CUISINE_TYPE_01_ID)))
+                .andExpect(jsonPath("$[0].name", is(CuisineTypesExamples.CUISINE_TYPE_01_NAME_SPANISH)));
     }
 
     @Test
     public void getAllCuisineTypesNoLanguage() throws Exception {
 
-        when(this.cuisineTypesBusiness.getCuisineType(any(), any())).thenReturn(Optional.of(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH));
+        when(this.cuisineTypesBusiness.getCuisineTypes(any(), any())).thenReturn(Arrays.asList(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH, CuisineTypesExamples.CUISINE_TYPE_BO_02_ENGLISH, CuisineTypesExamples.CUISINE_TYPE_BO_03_ENGLISH));
 
-        this.mvc.perform(get("/cuisine-types/{cuisine-type}", CuisineTypesExamples.CUISINE_TYPE_01_ID)
+        this.mvc.perform(get("/cuisine-types")
                 .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isBadRequest())
                 //TODO DMS Meter interceptor general para que ponga las cabeceras correctas
@@ -103,9 +121,9 @@ public class CuisineTypesApiControllerGetCuisineTypeTests {
     @Test
     public void getAllCuisineTypesInvalidLanguage() throws Exception {
 
-        when(this.cuisineTypesBusiness.getCuisineType(any(), any())).thenReturn(Optional.of(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH));
+        when(this.cuisineTypesBusiness.getCuisineTypes(any(), any())).thenReturn(Arrays.asList(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH, CuisineTypesExamples.CUISINE_TYPE_BO_02_ENGLISH, CuisineTypesExamples.CUISINE_TYPE_BO_03_ENGLISH));
 
-        this.mvc.perform(get("/cuisine-types/{cuisine-type}", CuisineTypesExamples.CUISINE_TYPE_01_ID)
+        this.mvc.perform(get("/cuisine-types")
                 .header(HttpHeaders.ACCEPT_LANGUAGE, INVALID_LANGUAGE)
                 .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isBadRequest())
@@ -120,30 +138,31 @@ public class CuisineTypesApiControllerGetCuisineTypeTests {
     }
 
     @Test
-    public void getAllCuisineTypesByIdNotFound() throws Exception {
+    public void getAllCuisineTypesByNameNotFound() throws Exception {
 
-        when(this.cuisineTypesBusiness.getCuisineType(eq(CuisineTypesExamples.CUISINE_TYPE_01_ID), any())).thenReturn(Optional.of(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH));
-        when(this.cuisineTypesBusiness.getCuisineType(eq(INVALID_ID), any())).thenReturn(Optional.empty());
+        when(this.cuisineTypesBusiness.getCuisineTypes(any(), any())).thenReturn(Arrays.asList(CuisineTypesExamples.CUISINE_TYPE_BO_01_ENGLISH, CuisineTypesExamples.CUISINE_TYPE_BO_02_ENGLISH, CuisineTypesExamples.CUISINE_TYPE_BO_03_ENGLISH));
+        when(this.cuisineTypesBusiness.getCuisineTypes(INVALID_NAME, LanguageEnumBO.EN)).thenReturn(new ArrayList<>());
 
-        this.mvc.perform(get("/cuisine-types/{cuisine-type}", INVALID_ID)
+        this.mvc.perform(get("/cuisine-types")
                 .header(HttpHeaders.ACCEPT_LANGUAGE, LOCALE_ENGLISH)
-                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNotFound())
-                .andExpect(jsonPath("$.code", is(messages.get("error.notfound.code"))))
-                .andExpect(jsonPath("$.description", is(messages.get("error.notfound.desc"))));
+                .accept(MediaType.APPLICATION_JSON)
+                .param("name", INVALID_NAME))
+                .andExpect(status().isOk())
+                .andExpect(header().string(HttpHeaders.CONTENT_LANGUAGE, LOCALE_ENGLISH))
+                .andExpect(jsonPath("$.length()", is(0)));
     }
 
     @Test
     public void getAllCuisineTypesTypesOutOfMemoryException() throws Exception {
 
-        when(this.cuisineTypesBusiness.getCuisineType(any(), any())).thenThrow(new OutOfMemoryError());
+        when(this.cuisineTypesBusiness.getCuisineTypes(any(), any())).thenThrow(new OutOfMemoryError());
 
-        this.mvc.perform(get("/cuisine-types/{cuisine-type}", INVALID_ID)
+        this.mvc.perform(get("/cuisine-types")
                 .header(HttpHeaders.ACCEPT_LANGUAGE, LOCALE_ENGLISH)
                 .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isInternalServerError())
                 .andExpect(jsonPath("$.code", is(messages.get("error.server.code"))))
-                .andExpect(jsonPath("$.description", is(messages.get("error.server.desc", LOCALE_ENGLISH))));
+                .andExpect(jsonPath("$.description", is(messages.get("error.server.desc"))));
     }
