+++ src/test/java/com/ojodev/cookinghero/recipes/api/CuisineTypesApiControllerPostTest.java
@@ -32,7 +32,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @AutoConfigureMockMvc
-public class CuisineTypesApiControllerPostTests {
+public class CuisineTypesApiControllerPostTest {
 
     @Autowired
     private MockMvc mvc;
