+++ src/test/java/com/ojodev/cookinghero/recipes/api/RecipesControllerGetRecipesTest.java
@@ -30,7 +30,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @AutoConfigureMockMvc
-public class RecipesControllerGetRecipesTests {
+public class RecipesControllerGetRecipesTest {
 
 	@Autowired
 	private MockMvc mvc;
