+++ src/test/java/com/ojodev/cookinghero/recipes/mapper/LanguageEnumMapperTest.java
@@ -15,7 +15,7 @@ import static org.junit.Assert.assertNotNull;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class LanguageEnumMapperTests {
+public class LanguageEnumMapperTest {
 
 
     @Autowired
